Port QRegion to use QAtomic instead of QShared. Done for all 4 platforms.  Added a shared_empty and removed isNull() with the idea that it wasn't really used anywhere correctly (aside from QVariant perhaps). Rejoined the X11 and embedded versions together since they are essentially the same thing.
I really hope I didn't forget anything.

[git-p4: depot-paths = "//depot/qt/main/": change = 105111]

+++ src/network/qdns.cpp
@@ -1370,7 +1370,7 @@ QPtrList<QDnsRR> * QDnsDomain::cached( const QDns * r )
 	    // wait - if it's an unqualified name, only ask when all
 	    // the other alternatives are exhausted.
 	    if ( q == m->queries.size() && ( s.find( '.' ) >= 0 ||
-					     l->count() >= n.count()-1 ) ) {
+					     (int)l->count() >= n.count()-1 ) ) {
 		QDnsQuery * query = new QDnsQuery;
 		query->started = now();
 		query->id = ++::id;
